+++ src/modules/html2pdf.ts
@@ -5,9 +5,8 @@
 import { existsSync } from "fs" // Import function to check if a file exists
 import { HtmlValidate } from "html-validate/node" // Import the HTML validation library
 import { fetchHtmlFromUrl, readHtmlFromFilePath } from "../utilies.js" // Import custom utility functions to fetch HTML from URL or file path
-import { Page } from "puppeteer"
 import type { Report } from "html-validate"
-import { firefox, launchBrowser } from "../browsers.js"
+import { closeBrowser, launchBrowser, getPage, restorePage } from "../browsers.js"
 
 export type Options = {
   /**
@@ -25,39 +24,11 @@ export type Options = {
    * @param {boolean} [validation] - Optional flag to enable HTML validation (default is true).
    */
   validation?: boolean
-}
-
-const pagePoolSize = 5
-const RESOURCE_LIMIT = 100 // Maximum allowed resources
-let resourceCount = 0
-let pagePool: Page [ ] = [ ]
-
-async function launchPages ( ) {
-  if ( pagePool.length > 0 ) {
-    return pagePool
-  }
-
-  pagePool = await Promise.all ( Array.from ( { length: pagePoolSize }, async ( ) => {
-    if ( firefox && firefox.connected ) {
-      const page = await firefox.newPage ( )
-      await page.setRequestInterception ( true )
-      await page.setDefaultNavigationTimeout ( 10000 ) // 10 seconds
-      await page.goto ( "about:blank" ) // Load a blank page
-      page.on ( "request", request => {
-        resourceCount++
-        if ( resourceCount > RESOURCE_LIMIT ) {
-          page.reload ( ) // Reload the page when limit is exceeded
-          resourceCount = 0 // Reset the counter
-        } else {
-          request.continue ( )
-        }
-      } )
-      return page
-    } else {
-      throw new Error ( "Browser not available" )
-    }
-  } ) )
-  return pagePool
+  /**
+   * - If true, the browser will be closed after conversion.
+   * @param {boolean} [closeBrowser] - Optional flag to close the browser after conversion. Default is false.
+   */
+  closeBrowser?: boolean
 }
 
 /**
@@ -71,7 +42,7 @@ export const html2pdf = async (
   input: string | URL, // HTML input as a string, URL, or file path
   options: Options = { } // Optional flag to return the PDF as a base64 string
 ) => {
-  await launchBrowser ( "firefox" ) // Ensure the Firefox browser is launched
+  const { browser, type } = await launchBrowser ( ) // Ensure the Firefox browser is launched
 
   const validator = new HtmlValidate ( options.rules ?? {
     extends: [ "html-validate:standard" ], // Use the standard HTML validation rules
@@ -91,18 +62,11 @@ export const html2pdf = async (
     htmlContent = await readHtmlFromFilePath ( htmlContent )
   }
 
-  if ( !firefox?.connected ) {
+  if ( !browser?.connected ) {
     throw new Error ( "Browser not available" )
   }
 
-  const pagePool = await launchPages ( )
-
-  let page = pagePool.pop ( )
-  let tempPage = false
-  if ( !page ) {
-    tempPage = true
-    page = await firefox.newPage ( )
-  }
+  const page = await getPage ( type )
 
   const validation = ( options.validation ?? true ) // Enable HTML validation by default
   try {
@@ -147,15 +111,9 @@ export const html2pdf = async (
       }
     }
   } finally {
-    if ( tempPage ) {
-      page.close ( )
-    } else {
-      await page.setViewport ( {
-        width: 800,
-        height: 600,
-        deviceScaleFactor: 1
-      } )
-      pagePool.push ( page )
+    await restorePage ( type, page )
+    if ( options.closeBrowser ) {
+      await closeBrowser ( ) // Close the browser if specified in options
     }
   }
 }